# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Egypt Advocates law firm API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lawyers/{id}:
    get:
      operationId: getLawyer
      tags: [lawyers]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lawyer" }
        "404": { $ref: "#/components/responses/NotFound" }
  /services:
    get:
      operationId: listServices
      tags: [services]
      parameters:
        - in: query
          name: practiceAreaId
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Service" }
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
        "404": { $ref: "#/components/responses/NotFound" }
  /appointments/availability:
    get:
      operationId: getAvailability
      tags: [appointments]
      parameters:
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
        - in: query
          name: serviceId
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AvailabilityResponse" }
  /appointments:
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book a new appointment (guest, no auth)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAppointmentInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /legal-articles:
    get:
      operationId: listLegalArticles
      tags: [legal-library]
      parameters:
        - in: query
          name: q
          required: false
          schema: { type: string }
        - in: query
          name: categoryId
          required: false
          schema: { type: integer }
        - in: query
          name: limit
          required: false
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          required: false
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LegalArticleList" }
  /legal-articles/{slug}:
    get:
      operationId: getLegalArticle
      tags: [legal-library]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LegalArticle" }
        "404": { $ref: "#/components/responses/NotFound" }
  /blog-posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      parameters:
        - in: query
          name: limit
          required: false
          schema: { type: integer, default: 12 }
        - in: query
          name: offset
          required: false
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BlogPost" }
  /blog-posts/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlogPost" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /chat/threads:
    post:
      operationId: createChatThread
      tags: [chat]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateChatThreadInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatThread" }
  /chat/threads/{id}:
    get:
      operationId: getChatThread
      tags: [chat]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatThreadWithMessages" }
        "404": { $ref: "#/components/responses/NotFound" }
  /chat/threads/{id}/messages:
    post:
      operationId: postChatMessage
      tags: [chat]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostChatMessageInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatPostResult" }

  # ---------------- Admin Auth ----------------
...
components:
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Lawyer:
      type: object
      required: [id, slug, nameAr, nameEn, titleAr, titleEn, bioAr, bioEn, photoUrl, email, phone, specializations, yearsExperience, sortOrder, isActive]
      properties:
        id: { type: integer }
        slug: { type: string }
        nameAr: { type: string }
        nameEn: { type: string }
        titleAr: { type: string }
        titleEn: { type: string }
        bioAr: { type: string }
        bioEn: { type: string }
        photoUrl: { type: ["string", "null"] }
        email: { type: string }
        phone: { type: string }
        specializations:
          type: array
          items: { type: string }
        yearsExperience: { type: integer }
        sortOrder: { type: integer }
        isActive: { type: boolean }
...
    Service:
      type: object
      required: [id, slug, nameAr, nameEn, descriptionAr, descriptionEn, durationMinutes, priceEgp, deliveryMode, isActive]
      properties:
        id: { type: integer }
        slug: { type: string }
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        durationMinutes: { type: integer }
        priceEgp: { type: number }
        deliveryMode: { type: string, enum: [online, in_office, both] }
        practiceAreaId: { type: ["integer", "null"] }
        isActive: { type: boolean }
...
    AvailabilityResponse:
      type: object
      required: [date, slots]
      properties:
        date: { type: string, format: date }
        slots:
          type: array
          items:
            type: object
            required: [time, available]
            properties:
              time: { type: string, description: "HH:MM" }
              available: { type: boolean }
    Appointment:
      type: object
      required: [id, clientName, clientEmail, clientPhone, serviceId, scheduledAt, durationMinutes, mode, status, paymentMethod, paymentStatus, amountEgp, language, createdAt]
      properties:
        id: { type: integer }
        clientName: { type: string }
        clientEmail: { type: string }
        clientPhone: { type: string }
        serviceId: { type: integer }
        serviceNameAr: { type: ["string", "null"] }
        serviceNameEn: { type: ["string", "null"] }
        lawyerId: { type: ["integer", "null"] }
        lawyerNameAr: { type: ["string", "null"] }
        lawyerNameEn: { type: ["string", "null"] }
        scheduledAt: { type: string, format: date-time }
        durationMinutes: { type: integer }
        mode: { type: string, enum: [online, in_office] }
        notes: { type: ["string", "null"] }
        status: { type: string, enum: [pending, approved, rejected, completed, cancelled] }
        meetingLink: { type: ["string", "null"] }
        paymentMethod: { type: string, enum: [instapay, vodafone_cash, fawry, visa, cash, bank_transfer] }
        paymentStatus: { type: string, enum: [pending, confirmed, failed, refunded] }
        paymentReference: { type: ["string", "null"] }
        amountEgp: { type: number }
        language: { type: string, enum: [ar, en] }
        createdAt: { type: string, format: date-time }
    CreateAppointmentInput:
      type: object
      required: [clientName, clientEmail, clientPhone, serviceId, scheduledAt, mode, paymentMethod, language]
      properties:
        clientName: { type: string }
        clientEmail: { type: string }
        clientPhone: { type: string }
        serviceId: { type: integer }
        lawyerId: { type: ["integer", "null"] }
        scheduledAt: { type: string, format: date-time }
        mode: { type: string, enum: [online, in_office] }
        notes: { type: ["string", "null"] }
        paymentMethod: { type: string, enum: [instapay, vodafone_cash, fawry, visa, cash, bank_transfer] }
        language: { type: string, enum: [ar, en] }
...
    LegalArticle:
      type: object
      required: [id, slug, categoryId, titleAr, titleEn, summaryAr, summaryEn, contentAr, contentEn, lawNumber, year, tags, publishedAt, isPublished, createdAt]
      properties:
        id: { type: integer }
        slug: { type: string }
        categoryId: { type: integer }
        categoryNameAr: { type: ["string", "null"] }
        categoryNameEn: { type: ["string", "null"] }
        titleAr: { type: string }
        titleEn: { type: string }
        summaryAr: { type: string }
        summaryEn: { type: string }
        contentAr: { type: string }
        contentEn: { type: string }
        lawNumber: { type: ["string", "null"] }
        year: { type: ["integer", "null"] }
        tags:
          type: array
          items: { type: string }
        publishedAt: { type: ["string", "null"], format: date-time }
        isPublished: { type: boolean }
        createdAt: { type: string, format: date-time }
    LegalArticleList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/LegalArticle" }
        total: { type: integer }
...
    BlogPost:
      type: object
      required: [id, slug, titleAr, titleEn, summaryAr, summaryEn, contentAr, contentEn, authorName, publishedAt, isPublished, tags, createdAt]
      properties:
        id: { type: integer }
        slug: { type: string }
        titleAr: { type: string }
        titleEn: { type: string }
        summaryAr: { type: string }
        summaryEn: { type: string }
        contentAr: { type: string }
        contentEn: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        authorName: { type: string }
        publishedAt: { type: ["string", "null"], format: date-time }
        isPublished: { type: boolean }
        tags:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
...
    ChatThread:
      type: object
      required: [id, visitorName, language, status, lastMessageAt, createdAt]
      properties:
        id: { type: integer }
        visitorName: { type: string }
        visitorEmail: { type: ["string", "null"] }
        language: { type: string, enum: [ar, en] }
        status: { type: string, enum: [open, awaiting_support, closed] }
        assignedTo: { type: ["string", "null"] }
        lastMessageAt: { type: string, format: date-time }
        unreadByAdmin: { type: integer }
        createdAt: { type: string, format: date-time }
    ChatMessage:
      type: object
      required: [id, threadId, senderType, senderName, content, createdAt]
      properties:
        id: { type: integer }
        threadId: { type: integer }
        senderType: { type: string, enum: [visitor, bot, agent] }
        senderName: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
    ChatThreadWithMessages:
      type: object
      required: [thread, messages]
      properties:
        thread: { $ref: "#/components/schemas/ChatThread" }
        messages:
          type: array
          items: { $ref: "#/components/schemas/ChatMessage" }
    CreateChatThreadInput:
      type: object
      required: [visitorName, language, initialMessage]
      properties:
        visitorName: { type: string }
        visitorEmail: { type: ["string", "null"] }
        language: { type: string, enum: [ar, en] }
        initialMessage: { type: string }
    PostChatMessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
    ChatPostResult:
      type: object
      required: [userMessage, replies, isSupportOnline]
      properties:
        userMessage: { $ref: "#/components/schemas/ChatMessage" }
        replies:
          type: array
          items: { $ref: "#/components/schemas/ChatMessage" }
        isSupportOnline: { type: boolean }
...